copy sync app tests to a new test file
